+++ submodules/LegacyComponents/PublicHeaders/LegacyComponents/TGLetteredAvatarView.h
@@ -9,7 +9,7 @@
 
 - (void)setTitleNeedsDisplay;
 
-- (void)loadUserPlaceholderWithSize:(CGSize)size uid:(int)uid firstName:(NSString *)firstName lastName:(NSString *)lastName placeholder:(UIImage *)placeholder;
+- (void)loadUserPlaceholderWithSize:(CGSize)size uid:(int64_t)uid firstName:(NSString *)firstName lastName:(NSString *)lastName placeholder:(UIImage *)placeholder;
 - (void)loadGroupPlaceholderWithSize:(CGSize)size conversationId:(int64_t)conversationId title:(NSString *)title placeholder:(UIImage *)placeholder;
 - (void)loadSavedMessagesWithSize:(CGSize)size placeholder:(UIImage *)placeholder;
