+++ submodules/LegacyComponents/PublicHeaders/LegacyComponents/TGMediaPickerCaptionInputPanel.h
@@ -22,7 +22,7 @@
 - (void)setCollapsed:(bool)collapsed animated:(bool)animated;
 
 - (void)replaceMention:(NSString *)mention;
-- (void)replaceMention:(NSString *)mention username:(bool)username userId:(int32_t)userId;
+- (void)replaceMention:(NSString *)mention username:(bool)username userId:(int64_t)userId;
 - (void)replaceHashtag:(NSString *)hashtag;
 
 - (void)adjustForOrientation:(UIInterfaceOrientation)orientation keyboardHeight:(CGFloat)keyboardHeight duration:(NSTimeInterval)duration animationCurve:(NSInteger)animationCurve;
